delete class thrimCli
